Move server startup logic into imports/startup/server/index.js
